move toggleElement to utility function file
