+++ src/components/ReservationForm.tsx
@@ -34,6 +34,7 @@ import Project from "../resources/Project";
 import Equipment from "../resources/Equipment";
 import { quantizeEquipment, buildDictionaries } from "../utils/equipment";
 import { findProjectById } from "../utils/project";
+import {toggleElement} from "../utils/div";
 
 const useStyles = makeStyles(() => ({
   guests: {
@@ -95,17 +96,6 @@ const ReservationForm: FunctionComponent<CalendarUIProps> = ({
       );
     } else {
       setValidationSchema(initialValidationSchema);
-    }
-  };
-  const toggleElement = (event: React.ChangeEvent<{}>, value: string): void => {
-    const val = (event.target as HTMLInputElement).value;
-    const element = document.getElementById(value);
-    if (element != null) {
-      if (val === "yes") {
-        element.style.display = "block";
-      } else {
-        element.style.display = "none";
-      }
     }
   };
   return (
